feat: implement IMU calibration function

+++ lib/Hack/taskGY521.h
@@ -3,14 +3,14 @@
 #define IMU_ADDRESS 0x68
 
 struct XyzData {
-  int x;
-  int y;
-  int z;
+  double x;
+  double y;
+  double z;
 };
 
 struct ImuData {
-  XyzData acceleration;
-  XyzData rotation;
+  XyzData acceleration;  // in m/s^2
+  XyzData rotation;      // in rad/s
   int temperature;
 };
